+++ src/index.ts
@@ -1,31 +1,18 @@
-import { ApolloServer } from "apollo-server-express";
-import app from "./express";
-import prisma from "./prisma";
+import express from "express";
+const app = express();
+import server from "./apollo";
 
-import { typedefs } from "./graphql/schema";
-import { gqlResolvers } from "./graphql/resolvers";
-
-import auth from "./services/auth";
-import events from "./services/events";
-import attendees from "./services/attendees"
-
-const server = new ApolloServer({
-  typeDefs: typedefs,
-  resolvers: gqlResolvers,
-  context: (req) => {
-    return {
-      ...req,
-      services: {
-        auth: auth,
-        events: events,
-        attendees: attendees
-      }
-    };
-  },
-});
-
+// apollo-server-middleware needs to be applied before express middleware
 server.applyMiddleware({ app });
 
-app.listen({ port: 4000 }, () =>
+import compression from "compression";
+import cookieParser from "cookie-parser";
+
+app.use(compression());
+app.use(cookieParser());
+
+app.listen({ port: process.env.PORT || 4000 }, () =>
   console.log(`🚀 Server ready at http://localhost:4000${server.graphqlPath}`)
 );
+
+export default app;
